+++ install.sh
@@ -17,11 +17,19 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+#######################################################################
+#                           GLOBAL VARIABLES                          #
+#######################################################################
+
 BUILDDIR="build"
 BINNAME="ukit"
 PREFIX="/"
 BINDIR="/usr/bin"
 MANDIR="/usr/share/man"
+
+#######################################################################
+#                           ENTRY POINT                               #
+#######################################################################
 
 args=$(getopt -a -n install -o p: --long prefix:,bindir:,mandir: -- "$@")
 [ $? -eq 1 ] && exit 1